Add the horizontal and vertical edge pieces.
Decided that I need an extra 1 cell margin around the whole thing to make
room for the wiring and electronics stuff, so the edge parts are longer
than the inner walls to accommodate that.

Vertically there will be no walls at all in the margin to make room for
the tangle of wires connecting together the LED strips. Horizontally there
will be a wall every 2 cells to support the grid while making room for
some of the larger electronics parts that don't fit inside 1 grid cell.