Make the game of life sound-reactive.
This is simple enough: we run the ADC in free-run mode and then do FFT,
and then any frequency above a threshold gets an R-pentomino drawn at
it. Since r-pentominos explode, this always causes something moderately
interesting to happen, but generally single frequencies don't arrive
in isolation so the explosions will often collide with one another in
interesting ways, causing a seemingly-organic growth that reacts to
the sound.