Apply a variable noise adjustment to each frequency column.
The ExtraCore-based circuit has a lot more noise than the Arduino one so
it's necessary to compensate for it when we're sampling, or else the
Game never stops animating.

This still seems a little off -- it's missing certain loud sounds and
still reacting to seemingly-quiet environments -- but I'll iterate on
the noise adjustments separately.